% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow.r
\name{square}
\alias{square}
\title{A square function}
\usage{
square(x)
}
\arguments{
\item{x}{a numeric value}
}
\value{
The square of x
}
\description{
takes a numeric value and squares it
}
\examples{
square(2)
square(3)
}
